+++ src/Electron/src/Swate.Electron.Shared/IPCTypes.fs
@@ -48,8 +48,11 @@ type IArcVaultsApi = {
 
     pickArcPaths: unit -> JS.Promise<Result<string[], exn>>
     pickDirectory: unit -> JS.Promise<Result<string, exn>>
-    pickAbsolutePaths: unit -> JS.Promise<Result<string[], exn>>
+    pickAbsolutePaths: unit -> JS.Promise<Result<string option, exn>>
     pickExternalTextFiles: unit -> JS.Promise<Result<ImportedTextFile[], exn>>
+    tryImportExternalFiles: ImportExternalFilesRequest -> JS.Promise<Result<ImportExternalFilesResult, exn>>
+    cancelImportExternalFiles: string -> JS.Promise<Result<unit, exn>>
+    getActiveFileImport: unit -> JS.Promise<Result<ActiveFileImportState option, exn>>
     getFileTree: unit -> JS.Promise<Result<System.Collections.Generic.Dictionary<string, FileEntry>, exn>>
     pathExists: string -> JS.Promise<Result<bool, exn>>
     openFile: string -> JS.Promise<Result<FileContentDTO, exn>>
@@ -152,6 +155,10 @@ module MainToRendererIpc =
 
     type IFileTreeRendererApi = {
         fileTreeUpdate: System.Collections.Generic.Dictionary<string, FileEntry> -> unit
+    }
+
+    type IFileImportRendererApi = {
+        fileImportStateUpdate: ActiveFileImportState option -> unit
     }
 
     type IGitProgressRendererApi = {